refactor(server): extract execution tools into src/server/handlers/execution-tools.ts
step_over / step_into / step_out (one ToolHandler, stepType still derived
from toolName exactly as before), continue_execution, pause_execution and
list_threads move off the CallTool switch; bodies verbatim with `this.` ->
`ctx.` and `result = X; break;` -> `return X;`. The try-terminal and
catch-branch-terminal assignments become returns (the case's break followed
the try/catch directly).

handlePause and handleListThreads move too; the class keeps one-line private
delegates as the test seams (removed in PR 6).
